chimera_Process

Chimera process CLIP data

chimera_Process(
  bams,
  knownMiRNAs,
  genomeIndex,
  exclude = NULL,
  bpparam = NULL,
  verbose = FALSE,
  removedups = TRUE,
  overwrite = FALSE
)

Arguments

bams
path to BAM files mapped to the genome, unmapped reads will be extracted for chimera processing.

knownMiRNAs
path to FASTA file containing annotated miRNA or other small RNA sequence. Known miRNAs will be prioritized and known miRNA names must be in the format "miR-", "let-", "bantam-", "iab-".

genomeIndex
path to genome index.

exclude
names of small RNAs to remove, default is NULL, can be set to character vector to specify; must match names in knownMiRNAs file.

bpparam
TRUE or FALSE (default).

verbose
print messages, TRUE or FALSE (default).

removedups
remove multiple small RNAs mapping to the same read, TRUE (default) or FALSE. If TRUE, known miRNAs will be prioritized and known miRNA names must be in the format "miR-", "let-", "bantam-", "iab-".

overwrite
overwrite existing output files, TRUE or FALSE (default).

Value

path to chimera output table.

Author

Kathryn Rozen-Gagnon

Examples

testFastq <- system.file("extdata/SRR1742056.fastq.gz",package="CLIPflexR")
FqFile <- decompress(testFastq,overwrite=TRUE)
FaFile <- fastx_qtoa(FqFile)
FaFile_clip <- fastx_clipper(FaFile, writelog=FALSE)
myGenome <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <- bowtie2_index(myGenome, overwrite = TRUE)

#> Warning: For argument `bt2Index`, file exist:/Users/runner/work/_temp/Library/CLIPflexR/extdata/hg19Small.1.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/Users/runner/work/_temp/Library/CLIPflexR/extdata/hg19Small.2.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/Users/runner/work/_temp/Library/CLIPflexR/extdata/hg19Small.3.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/Users/runner/work/_temp/Library/CLIPflexR/extdata/hg19Small.4.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/Users/runner/work/_temp/Library/CLIPflexR/extdata/hg19Small.rev.1.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/Users/runner/work/_temp/Library/CLIPflexR/extdata/hg19Small.rev.2.bt2. It will be overwrited

myBam <- bowtie_align(FaFile_clip,myIndex, overwrite=TRUE)
miRNAs <- system.file("extdata/hsa_mature.fa",package="CLIPflexR")
chimera_bed <- chimera_Process(myBam, miRNAs, myIndex, exclude="hsa-miR-19a-3p", overwrite=TRUE)

#> Warning, the following sample(s) have duplicate read names!