Wrapper function for ctk's ctk_tag2profile

Wrapper function for ctk's ctk_tag2profile

ctk_tag2profile(
  filesToRun,
  outFile = paste0(file_path_sans_ext(filesToRun), ".out"),
  outFile2 = NULL,
  sb = "tag2profile.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  bigFile = FALSE,
  weight = FALSE,
  weightAvg = FALSE,
  ss = TRUE,
  exact = TRUE,
  nz = FALSE,
  ext5 = NULL,
  ext3 = NULL,
  chromLen = NULL,
  region = NULL,
  minBlockSize = 2e+06,
  windowSize = 100,
  stepSize = 20,
  outputFormat = "bedgraph",
  normalization = "none",
  stderr = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_tag2profile_stderr.txt")),
  stdout = file.path(dirname(fileToRun), paste0(basename(fileToRun),
    "_ctk_tag2profile_stdout.txt")),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additional_Args = NULL,
  verbose = FALSE,
  writelog = T
)

Arguments

filesToRun
path to file to process (BED).

outFile
path to output file (BED).

outFile2
path to output 2 file (BED; specify two output files to separate strands).

sb
path to tag2profile.pl from CTK.

perl
path to PERL

PATHTOPERLLIB
path to PERL5LIB.

bigFile
TRUE when working with a big file, TRUE or FALSE (default).

weight
weight counts according to the score of each tag, TRUE or FALSE (default).

weightAvg
weight average the score of each tag, TRUE or FALSE (default).

ss
separate strand, TRUE (default) or FALSE.

exact
exact count at each nucleotide, TRUE (default) or FALSE.

nz
don't print zeroes (works for sgr and bed), TRUE or FALSE (default).

ext5
extension of tags at the 5' end, default is NULL (set to integer to specify).

ext3
extension of tags at the 3' end, default is NULL (set to integer to specify).

chromLen
chrom length file, default is NULL (set file path to specify).

region
a bed file with regions to count tag numbers; if not specified (NULL), count in moving windows

minBlockSize
minimum number of lines to read in each block for a big file, default is 2000000.

windowSize
window size, default is 100.

stepSize
step size, default is 20.

outputFormat
output format, "bed" or "bedgraph" (default) or "sgr".

normalization
normalization, "none" (default) or "rpkm" or multiply=1.3).

stderr
path to stdout file.

stdout
path to stdout file.

useClipRConda
use conda environment installed by Herper, TRUE (default) or FALSE.

additional_Args
Additional arguments to be passed to system call.

verbose
print messages, TRUE or FALSE (default).

writelog
write stderr/stdout logs, TRUE (default) or FALSE.

Value

path to BED file.

Examples

testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <-suppressWarnings(bowtie2_index(testFasta, overwrite = TRUE))
testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter="mean:0-29:20",verbose=TRUE)

#> fastq_filter.pl command is /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.20/bin/ctk/fastq_filter.pl

#> fastq_filter.pl arguments are /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.20/bin/ctk/fastq_filter.pl  -if sanger  -of fastq  -f mean:0-29:20  /Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/Fox3_Std_small.fq.gz  /Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/FF_Fox3_Std_small.fq.gz 

FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)

#> fastq_filter.pl command is /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.20/bin/ctk/fastq2collapse.pl

#> fastq_filter.pl arguments are /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.20/bin/ctk/fastq2collapse.pl  /Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/QT_FF_Fox3_Std_small_clip.fq  /Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/Collapsed_QT_FF_Fox3_Std_small_clip.fq 

FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5,inputFormat="fastq")
bam <- bowtie_align(FqFile_QFColStripped,myIndex, overwrite=TRUE, inputFormat="fastq")

#> Warning: For argument `samOutput`, file exist:/Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/Collapsed_QT_FF_Fox3_Std_small_clip_rm5.sam. It will be overwrited

parsedAlignment <- ctk_parseAlignment(bam)
myCollaped <- ctk_tag2collapse(parsedAlignment,weight=FALSE,randomBarcode=FALSE,
weightInName=FALSE,verbose=TRUE)

#> tag2collapse.pl command is /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.20/bin/ctk/tag2collapse.pl

#> tag2collapse.pl arguments are /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.20/bin/ctk/tag2collapse.pl  --keep-tag-name   --keep-max-score   -EM -1  --seq-error-model TRUE  /Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/Collapsed_QT_FF_Fox3_Std_small_clip_rm5.bed  /Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/TC_Collapsed_QT_FF_Fox3_Std_small_clip_rm5.bed 

myrgbBed <- ctk_bed2rgb(myCollaped,col="128,0,0")
ctk_tag2profile(myrgbBed,verbose=TRUE)

#> tag2profile.pl command is /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.20/bin/ctk/tag2profile.pl

#> tag2profile.pl arguments are /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.20/bin/ctk/tag2profile.pl  -ss   -exact   -minBlockSize 2000000  -w 100  -s 20  -of bedgraph  -normalize none  /Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/TC_Collapsed_QT_FF_Fox3_Std_small_clip_rm5.RGB.bed  /Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/TC_Collapsed_QT_FF_Fox3_Std_small_clip_rm5.RGB.out 

#> [1] "/Library/Frameworks/R.framework/Versions/3.5/Resources/library/CLIPflexR/extdata/TC_Collapsed_QT_FF_Fox3_Std_small_clip_rm5.RGB.out"