Wrapper function for ctk's bed2rgb

Wrapper function for ctk's bed2rgb

ctk_tag2profile(
  filesToRun,
  outFile = paste0(file_path_sans_ext(filesToRun), ".out"),
  outFile2 = NULL,
  sb = "tag2profile.pl",
  perl = "perl",
  PATHTOPERLLIB = NULL,
  bigFile = FALSE,
  weight = FALSE,
  weightAvg = FALSE,
  ss = TRUE,
  exact = TRUE,
  nz = FALSE,
  ext5 = NULL,
  ext3 = NULL,
  chromLen = NULL,
  region = NULL,
  minBlockSize = 2e+06,
  windowSize = 100,
  stepSize = 20,
  outputFormat = "bedgraph",
  normalization = "none",
  stderr = paste0(getwd(), "stripBarcode_stderr"),
  stdout = paste0(getwd(), "stripBarcode_stdout"),
  useClipRConda = ifelse(is.null(getOption("CLIPflexR.condaEnv")), FALSE, TRUE),
  additionalArgumements = NULL,
  verbose = FALSE
)

Arguments

filesToRun
File to process.

outFile
Output file

outFile2
Output 2 file

sb
Path to tag2collapse.pl from FastX toolkit

perl
Path to PERL

PATHTOPERLLIB
Path to PERL5LIB.

bigFile
TRUE when working with a big file

weight
weight counts according to the score of each tag

weightAvg
weight average the score of each tag

ss
separate strand

exact
exact count at each nucleotide

nz
don't print zeroes (works for sgr and bed)

ext5
extension of tags at the 5' end

ext3
extension of tags at the 3' end

chromLen
chrom length file

region
a bed file with regions to count tag numbers. If not specified, count in moving windows

minBlockSize
minimum number of lines to read in each block for a big file

windowSize
Window size

stepSize
Step size

outputFormat
output format ([bed] or bedgraph or sgr)

normalization
normalization ([none] or rpkm or multiply=1.3)

stderr
Path to stdout file.

stdout
Path to stdout file.

useClipRConda
Boolean on whether to use conda environment install by CondaSysReqs

additionalArgumements
Additional arguments to be passed to system call.

verbose
Print more message to screen.

Value

Path to unzipped file

Examples

testFasta <- system.file("extdata/hg19Small.fa",package="CLIPflexR")
myIndex <- bowtie2_index(testFasta)

#> Warning: For argument `bt2Index`, file exist:/private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/hg19Small.1.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/hg19Small.2.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/hg19Small.3.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/hg19Small.4.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/hg19Small.rev.1.bt2. It will be overwrited

#> Warning: For argument `bt2Index`, file exist:/private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/hg19Small.rev.2.bt2. It will be overwrited

testFQ <- system.file("extdata/Fox3_Std_small.fq.gz",package="CLIPflexR")
FqFile_FF <- ctk_fastqFilter(testFQ,qsFilter = "mean:0-29:20",verbose=TRUE)

#> fastq_filter.pl command is /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.18/bin/ctk/fastq_filter.pl

#> fastq_filter.pl arguments are /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.18/bin/ctk/fastq_filter.pl  -if sanger  -of fastq  -f mean:0-29:20  /private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/Fox3_Std_small.fq.gz  /private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/FF_Fox3_Std_small.fq.gz 

FqFile <- decompress(FqFile_FF,overwrite=TRUE)
FqFile_clipped <- fastx_clipper(FqFile,length=20)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped)
FqFile_QF <- fastq_quality_trimmer(FqFile_clipped,paste0(FqFile_QF,".gz"))
FqFile_Col <- ctk_fastq2collapse(FqFile_QF,verbose=TRUE)

#> fastq_filter.pl command is /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.18/bin/ctk/fastq2collapse.pl

#> fastq_filter.pl arguments are /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.18/bin/ctk/fastq2collapse.pl  /private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/QF_FF_Fox3_Std_small_clip.fq.gz  /private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/Collapsed_QF_FF_Fox3_Std_small_clip.fq 

FqFile_QFColStripped <- ctk_stripBarcode(FqFile_Col,linkerlength=5)
bam <- bowtie_align(FqFile_QFColStripped,myIndex)
parsedAlignment <- ctk_parseAlignment(bam)
myCollaped <- ctk_tag2collapse(parsedAlignment,weight=FALSE,randomBarcode=FALSE,
weightInName=FALSE,verbose=TRUE)

#> fastq_filter.pl command is /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.18/bin/ctk/tag2collapse.pl

#> fastq_filter.pl arguments are /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.18/bin/ctk/tag2collapse.pl  --keep-max-score   --keep-tag-name   -EM -1  -big    /private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/Sorted_Collapsed_QF_FF_Fox3_Std_small_clip.fq_rm5.gzfa.bed  /private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/TC_Sorted_Collapsed_QF_FF_Fox3_Std_small_clip.fq_rm5.gzfa.bed 

myrgbBed <- ctk_bed2rgb(myCollaped,col="128,0,0")
ctk_tag2profile(myrgbBed,verbose=TRUE)

#> fastq_filter.pl command is /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.18/bin/ctk/tag2profile.pl

#> fastq_filter.pl arguments are /Users/kathryn/Library/r-miniconda/envs/CLIPflexR_0.1.18/bin/ctk/tag2profile.pl  -ss   -exact   -minBlockSize 2000000  -w 100  -s 20  -of bedgraph  -normalize none  /private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/TC_Sorted_Collapsed_QF_FF_Fox3_Std_small_clip.fq_rm5.gzfa.RGB.bed  /private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/TC_Sorted_Collapsed_QF_FF_Fox3_Std_small_clip.fq_rm5.gzfa.RGB.out 

#> [1] "/private/var/folders/rx/rt8glh297vbg70g644dpwy6w0000gn/T/Rtmp3FK8Ma/temp_libpathbc4c67bc603c/CLIPflexR/extdata/TC_Sorted_Collapsed_QF_FF_Fox3_Std_small_clip.fq_rm5.gzfa.RGB.out"